backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
PyPDF2==3.0.1
nltk==3.9.1
python-dotenv==1.0.1
flask
flask-cors
pymongo
python-dotenv
PyPDF2
python-docx
bcrypt
pyjwt
ollama
groq
reportlab
